[FEATURE] Support filtering on name and value fields in issue object fields
The issue API doesn't have a `value` field for the `project` key -
instead, it has a `name`. With this change, we check for both `value`
and `name` when looking for a field on an issue to match on.

There's some cleanup to the field lookup method to split it into a few
helpers, because it would have gotten quite complex with this change
without it.

LRN-12584